# Improved the performance of the DataReaderEntityMapper.