This PR includes two main changes. When running benchmarks I noticed that we were slow in two areas.

**Selecting a single record**
To improve performance of selecting a single record we store the EntityMapperDefinition in a static list. This definition contains the data types of each of the columns, the compiled lambda property setters, and the object activator. This took TOP 1 select on my benchmark from approx 1.5ms to 0.27ms.

This means first doesn't get any of the performance benefits, but I believe Dapper does a similar thing.

**Selecting a large amount of records**
To improve performance of selecting large amounts of records I have added a new Func<> which is created once for each column of the first record. For the rest of the rows we do not need to figure out what columns we have, or what the data types are. We can just pass our newly created Entity and the Data Reader to our Funcs which are already created. 

I have also removed the IsDBNull check from the non-nullable columns. 

This took selecting 250k records from 70ms to 43ms.