@MarkioE 2018-02-27
Component | TestName | MinTimeTakenMs
Dapper | SelectAllBenchmarkOperationDapper | 46.37
SqlRepo | SelectAllBenchmarkOperationSqlRepo | 43.47
Dapper | SelectByDecimalValueBenchmarkOperationDapper | 4.81
SqlRepo | SelectByDecimalValueBenchmarkOperationSqlRepo | 4.98
Dapper | SelectSingleColumnBenchmarkDapper | 15.34
SqlRepo | SelectSingleColumnBenchmarkSqlRepo | 14.71
Dapper | SelectTop1BenchmarkOperationDapper | 0.22
SqlRepo | SelectTop1BenchmarkOperationSqlRepo | 0.27
Dapper | SelectTop5000BenchmarkOperationDapper | 4.04
SqlRepo | SelectTop5000BenchmarkOperationSqlRepo | 3.9
Dapper | SelectWhereBetweenBenchmarkOperationDapper | 5.96
SqlRepo | SelectWhereBetweenBenchmarkOperationSqlRepo | 6.11